+++ BoardGameReviewsBackend.API/Requests/Boardgames/GetBoardgameRequest.cs
@@ -2,5 +2,5 @@ namespace BoardGameReviewsBackend.API.Requests.Boardgames;
 
 public class GetBoardgameRequest
 {
-    public long boardgameId { get; set; }
+    public int boardgameId { get; set; }
 }
